+++ Sources/ProcessOut/Sources/Repositories/CustomerTokens/Responses/PONativeAlternativePaymentTokenizationResponseV2.swift
@@ -14,8 +14,8 @@ public struct PONativeAlternativePaymentTokenizationResponseV2: Sendable, Decoda
     public let state: PONativeAlternativePaymentStateV2
 
     /// Next step if any.
-    public let nextStep: PONativeAlternativePaymentNextStepV2?
+    public let elements: [PONativeAlternativePaymentElementV2]?
 
-    /// Instructions providing additional information to customer and/or describing additional actions.
-    public let customerInstructions: [PONativeAlternativePaymentCustomerInstructionV2]?
+    /// Redirect details.
+    public let redirect: PONativeAlternativePaymentRedirectV2?
 }
